feat(server-info): add dash alias and multi-page for roles
